+++ docx-core/src/reader/attributes/indent.rs
@@ -8,21 +8,21 @@ use super::super::errors::*;
 
 pub fn read_indent(
     attrs: &[OwnedAttribute],
-) -> Result<(usize, Option<usize>, Option<SpecialIndentType>), ReaderError> {
+) -> Result<(i32, Option<i32>, Option<SpecialIndentType>), ReaderError> {
     let mut start = 0;
-    let mut end: Option<usize> = None;
+    let mut end: Option<i32> = None;
     let mut special: Option<SpecialIndentType> = None;
 
     for a in attrs {
         let local_name = &a.name.local_name;
         if local_name == "left" || local_name == "start" {
-            start = usize::from_str(&a.value)?;
+            start = i32::from_str(&a.value)?;
         } else if local_name == "end" || local_name == "right" {
-            end = Some(usize::from_str(&a.value)?);
+            end = Some(i32::from_str(&a.value)?);
         } else if local_name == "hanging" {
-            special = Some(SpecialIndentType::Hanging(usize::from_str(&a.value)?))
+            special = Some(SpecialIndentType::Hanging(i32::from_str(&a.value)?))
         } else if local_name == "firstLine" {
-            special = Some(SpecialIndentType::FirstLine(usize::from_str(&a.value)?))
+            special = Some(SpecialIndentType::FirstLine(i32::from_str(&a.value)?))
         }
     }
